+++ reg.h
@@ -11,13 +11,13 @@ int8_t register32_clear_bits(volatile uint32_t* reg, uint32_t mask);
 
 uint32_t register32_read(volatile uint32_t* reg);
 
-int8_t register16_write(volatile uint32_t* reg, uint16_t value);
+int8_t register16_write(volatile uint16_t* reg, uint16_t value);
 
-int8_t register16_set_bits(volatile uint32_t* reg, uint16_t mask);
+int8_t register16_set_bits(volatile uint16_t* reg, uint16_t mask);
 
-int8_t register16_clear_bits(volatile uint32_t* reg, uint16_t mask);
+int8_t register16_clear_bits(volatile uint16_t* reg, uint16_t mask);
 
-uint16_t register16_read(volatile uint32_t* reg);
+uint16_t register16_read(volatile uint16_t* reg);
 
 int8_t register8_write(volatile uint8_t* reg, uint8_t value);
